Separe managers from models
